In GUI Reserverd the word "Channel" for physical sensor channels
Reserved the word "Channel" in the GUI to refer to the "electrical" channels within the sensor.
Now use a (column,row) nomenclature for the physical arrangement of photoreceiver voxels.
This is to avoid confision with ADI documentation, which uses "channels" for chip channels, which may not be in order of the pixels.
Also preparing to better support 2D voxel array arrangements.